refactor: Identification Result Page
