+++ frontend-hybrid/yakal/lib/main.dart
@@ -4,13 +4,13 @@ import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:get/get.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:yakal/screens/Home/home_screen.dart';
+import 'package:yakal/screens/Login/Identification/screen.dart';
 import 'package:yakal/screens/Login/IdentificationEntry/screen.dart';
+import 'package:yakal/screens/Login/IdentificationResult/screen.dart';
 import 'package:yakal/screens/Login/LoginEntry/screen.dart';
 import 'package:yakal/screens/Login/LoginTerms/screen.dart';
 import 'package:yakal/screens/Login/auth_finish_screen.dart';
 import 'package:yakal/screens/Login/enter_nickname_screen.dart';
-import 'package:yakal/screens/Login/identification_result_screen.dart';
-import 'package:yakal/screens/Login/identification_screen.dart';
 import 'package:yakal/screens/Login/select_mode_screen.dart';
 import 'package:yakal/screens/Profile/profile_screen.dart';
 import 'package:yakal/screens/Setting/setting_screen.dart';
